individual project page extracted from project component and added as sperate page - scroll reset to top issue fixed, mobile navbar border color changed to grayish
